#114 Phase A: extract JS res.* helpers -> bindings_response.c (mirror)
Mirror of the Lua res:* extraction on the JS side: move the response cluster
(hl_response_class + get_response + hl_response_has_header + the js_res_*
handlers + hl_js_ensure_response_class + the public hl_js_make_response) out of
bindings.c into js/bindings_response.c, so the core js_bindings.o holds ZERO
Keel-response / compress references. Cleaner than the Lua move - the JS helpers
get HlJS via JS_GetContextOpaque and js_get_buffer is already shared in
mod_buffer.h, so no static helper needed exposing.

Auto-wired via JS_RT_OBJS; Phase C relocates it into the composed `http`
feature. Verified: `make test` 60/60; e2e_feature_runtime green (the js
res.json app serves); `nm` confirms js_bindings.o has no kl_response_* /
hl_maybe_compress refs.
